feat: add cursor pagination on post comments
Added PostCommentLog model to keep track of a comment's score
at a point in time. The reason for this is to create a stable cursor
when sorting comments by score. It solves the problem of duplicate
and skipped records. The problem is also mentioned in the link below:
https://stackoverflow.com/questions/78487292/how-can-i-implement-cursor-pagination-when-sorting-by-dynamic-attributes-such-as

When fetching comments sorted by score, the order of the comments are
based on the comment logs. There is at least 1 comment log per comment
and only the newest comment log is kept. Older ones are set to expire
in a day to free up space. So, if a user decides to scroll through
comments for more than a day, then some comments may be skipped.

This solution is probably horrible and overly complicated. I just wanted
to find another way to solve the problem of duplicate and skipped records
other than the simpler solution of storing comment ids in the frontend
and then sending them to the backend so that these ids can be excluded
when fetching for comments sorted by score (or any other criteria).
Hopefully, I can find better solutions in the future.